+++ SurgSim/Graphics/Camera.h
@@ -108,10 +108,10 @@ public:
 
 	/// Sets the width and height of the viewport
 	/// \param dimensions size of the viewport in screen sapce
-	virtual void setViewportSize(std::array<int, 2> dimensions) = 0;
+	virtual void setViewportSize(std::array<double, 2> dimensions) = 0;
 
 	/// Gets the dimensions of the viewport
-	virtual std::array<int, 2> getViewportSize() const = 0;
+	virtual std::array<double, 2> getViewportSize() const = 0;
 
 	/// Set the projection matrix with the appropriate  perspective projection parameters
 	/// \param fovy Field of view along the y-axis
